* commit '134aec026925c54226637ac2cb68fd049f15b67f':
  SW-22485 - Fix benchmark translations
  SW-22485 - correct BenchmarkBundle and SearchBundle phpdoc and incorrect types fixes #1792

+++ engine/Shopware/Bundle/BenchmarkBundle/Controllers/Backend/BenchmarkLocalOverview.php
@@ -113,7 +113,7 @@ class Shopware_Controllers_Backend_BenchmarkLocalOverview extends Shopware_Contr
         $now = new DateTime('now');
         $diff = $now->diff($this->getWaitingSinceDate());
 
-        return $diff->h + ($diff->days * 24);
+        return (int) ($diff->h + ($diff->days * 24));
     }
 
     /**
